fix: #WB-3001 refacto reaction components to use store + clean PostPreview component
