+++ frontend/src/features/Blog/BlogPostList.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 
-import { Button, EmptyScreen, usePaths, useReactions } from "@edifice-ui/react";
+import { Button, EmptyScreen, usePaths } from "@edifice-ui/react";
 import { useTranslation } from "react-i18next";
 
 import { PostPreview } from "~/components/PostPreview/PostPreview";
@@ -30,10 +30,7 @@ const BlogPostList = () => {
   const { postsFilters } = usePostsFilter();
   const { counters } = useBlogCounter(blog?._id);
 
-  const { sidebarHighlightedPost, postsViewsCounters, postsReactionsSummary } =
-    useBlogState();
-
-  const { availableReactions } = useReactions("blog", "post");
+  const { sidebarHighlightedPost } = useBlogState();
 
   useEffect(() => {
     if (sidebarHighlightedPost) {
@@ -86,20 +83,9 @@ const BlogPostList = () => {
           />
         </div>
       )}
-      {posts?.map((post, index) =>
-        postsReactionsSummary ? (
-          <PostPreview
-            key={post._id}
-            post={post}
-            index={index}
-            views={postsViewsCounters?.[post._id]}
-            reactions={{
-              available: availableReactions,
-              summary: postsReactionsSummary[post._id],
-            }}
-          />
-        ) : null,
-      )}
+      {posts?.map((post, index) => (
+        <PostPreview key={post._id} post={post} index={index} />
+      ))}
       {hasNextPage && (
         <div className="d-flex justify-content-center">
           <Button
